修复一些遗留bug，部分适配自定义标题栏

+++ CraftNow/src/CraftNow/Utils/PlatformUtils.h
@@ -16,6 +16,6 @@ namespace CraftNow {
 	class Time
 	{
 	public:
-		static float GetTime();
+		static double GetTime();
 	};
 }
